project/requirements.txt
django
psycopg2-binary
djangorestframework
django-extensions
requests
celery
redis
ipython
tenacity
flake8